+++ app/src/main/java/com/nextinnovation/pitak/model/car/Car.java
@@ -1,19 +1,19 @@
 package com.nextinnovation.pitak.model.car;
 
 public class Car {
-    private int id;
+    private long id;
     private String name;
 
-    public Car(int id, String name) {
+    public Car(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
